Move undoMachine into subfolder
